pyproject.toml
[project]
name = "whatsapp-agno"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.0.10",
    "celery>=5.5.3",
    "fastapi[standard]>=0.117.1",
    "openai>=1.109.1",
    "python-dotenv>=1.1.1",
    "redisvl>=0.8.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.1.1",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
]
